+++ browser/data-browser/src/components/forms/SearchBox/SearchBoxButton.tsx
@@ -1,4 +1,5 @@
 import { styled } from 'styled-components';
+import { SB_BOTTOM_RADIUS, SB_HIGHLIGHT, SB_TOP_RADIUS } from './searchboxVars';
 
 export const SearchBoxButton = styled.button<{ ephimeral?: boolean }>`
   background-color: transparent;
@@ -10,15 +11,27 @@ export const SearchBoxButton = styled.button<{ ephimeral?: boolean }>`
   padding: 0.5rem;
   color: ${p => p.theme.colors.textLight};
   cursor: pointer;
+  visibility: ${p => (p.ephimeral ? 'hidden' : 'visible')};
 
-  &:hover,
-  &:focus-visible {
-    color: var(--search-box-hightlight);
-    background-color: ${p => p.theme.colors.bg1};
-    border-color: var(--search-box-hightlight);
+  &:last-child {
+    border-top-right-radius: ${p => SB_TOP_RADIUS.var(p.theme.radius)};
+    border-bottom-right-radius: ${p => SB_BOTTOM_RADIUS.var(p.theme.radius)};
   }
 
-  visibility: ${p => (p.ephimeral ? 'hidden' : 'visible')};
+  &:disabled {
+    color: ${p => p.theme.colors.textLight2};
+    cursor: not-allowed;
+  }
+
+  &:not(:disabled) {
+    &:hover,
+    &:focus-visible {
+      color: ${SB_HIGHLIGHT.var()};
+      background-color: ${p => p.theme.colors.bg1};
+      border-color: ${SB_HIGHLIGHT.var()};
+    }
+  }
+
   div:hover > & {
     visibility: visible;
   }