+++ browser/data-browser/src/views/OntologyPage/PropertyDatatypePicker.tsx
@@ -46,6 +46,7 @@ export function PropertyDatatypePicker({
 
   return (
     <StyledAtomicSelectInput
+      aria-label='Property datatype'
       commit
       disabled={disabled}
       resource={resource}