+++ browser/CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog covers all five packages, as they are (for now) updated as a whol
 ### Atomic Browser
 
 - [#981](https://github.com/atomicdata-dev/atomic-server/issues/981) Fix bug where the service worker would not update cache with updated code.
+- [#989](https://github.com/atomicdata-dev/atomic-server/issues/989) Added an edit button to the resource selector inputs.
 
 ### @tomic/cli